1. finished endpoint for routes 2. db logic was moved to new module service(web) 3. some refactor
