--- web/src/main/java/com/vertex/config/Application.java
@@ -1,16 +1,0 @@
-package com.vertex.config;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.liquibase.LiquibaseAutoConfiguration;
-import org.springframework.context.ApplicationContext;
-
-/**
- * Created by RASTA on 25.05.2016.
- */
-@SpringBootApplication(exclude = LiquibaseAutoConfiguration.class)
-public class Application {
-    public static void main(String[] args) {
-        ApplicationContext context = SpringApplication.run(Application.class, args);
-    }
-}
